# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalshi Scalping Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotConfig {
  /** Only enter if ask price per contract is at or below this (cents) */
  maxEntryPriceCents: number;
  /** Minimum net profit to exit (cents, after fees) */
  minNetProfitCents: number;
  /** Maximum net profit target (cents, after fees) */
  maxNetProfitCents: number;
  /** Skip market if this many minutes or fewer remain */
  minMinutesRemaining: number;
  /** Kalshi fee rate on profit (e.g. 0.07 = 7%) */
  feeRate: number;
  /** How often to scan markets (seconds) */
  pollIntervalSecs: number;
}

export interface BotConfigUpdate {
  maxEntryPriceCents?: number;
  minNetProfitCents?: number;
  maxNetProfitCents?: number;
  minMinutesRemaining?: number;
  feeRate?: number;
  pollIntervalSecs?: number;
}

export interface BotStatus {
  running: boolean;
  /** @nullable */
  startedAt: string | null;
  marketsScanned: number;
  tradesAttempted: number;
  tradesSucceeded: number;
  totalPnlCents: number;
}

export interface Trade {
  id: number;
  marketId: string;
  marketTitle: string;
  side: string;
  buyPriceCents: number;
  /** @nullable */
  sellPriceCents: number | null;
  contractCount: number;
  feeCents: number;
  /** @nullable */
  pnlCents: number | null;
  status: string;
  createdAt: string;
  /** @nullable */
  closedAt: string | null;
  /** @nullable */
  minutesRemaining: number | null;
}

export interface TradeList {
  trades: Trade[];
  total: number;
}

export interface TradeStats {
  totalTrades: number;
  winningTrades: number;
  losingTrades: number;
  openTrades: number;
  totalPnlCents: number;
  winRate: number;
  avgPnlCents: number;
}

export interface LogEntry {
  id: number;
  level: string;
  message: string;
  /** @nullable */
  data: string | null;
  createdAt: string;
}

export interface LogList {
  logs: LogEntry[];
}

export type ListTradesParams = {
  limit?: number;
  offset?: number;
};

export type GetBotLogsParams = {
  limit?: number;
};