# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalshi Scalping Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfig {
  /** Only enter if ask price per contract is at or below this (cents) */
  maxEntryPriceCents: number;
  /** Minimum net profit to exit (cents, after fees) */
  minNetProfitCents: number;
  /** Maximum net profit target (cents, after fees) */
  maxNetProfitCents: number;
  /** Skip market if this many minutes or fewer remain */
  minMinutesRemaining: number;
  /** Kalshi fee rate on profit (e.g. 0.07 = 7%) */
  feeRate: number;
  /** How often to scan markets (seconds) */
  pollIntervalSecs: number;
}