# lib/api-zod/src/generated/types/botConfigUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalshi Scalping Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfigUpdate {
  maxEntryPriceCents?: number;
  minNetProfitCents?: number;
  maxNetProfitCents?: number;
  minMinutesRemaining?: number;
  feeRate?: number;
  pollIntervalSecs?: number;
}